mlops/requirements.txt
pandas
numpy
pymongo
ipykernel
matplotlib
plotly
seaborn
scipy
scikit-learn
imbalanced-learn
xgboost
catboost
from-root
evidently==0.2.8
dill
PyYAML
neuro-mf
boto3
mypy-boto3-s3
botocore
fastapi
